requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
google-generativeai
tavily-python
groq
